cluster picking: lock all the lines before set destination
At the beginning of the endpoint that writes the destination of lines
moved by a package, lock them. This is mostly to have an entrypoint for
the module shopfloor_checkout_sync:

In the general scenario, the lines are moved by the same bin package,
they should be moved by the same operator, so we should never have any
concurrency issue.

However, when we use the synchronization of destinations
(shopfloor_checkout_sync), as soon as we scan a destination for the
first line, it updates the destination of all the other move lines
that reach the same operation type for an order, so it might change
the destination of another operator's line. At the beginning of the
endpoint, lock the line to ensure 2 operators won't synchronize
different destinations (I would expect to have a transaction rollback
error and a rollback anyway).