front: load demo resources explicitily
Before this change demo resources were loaded automatically by the mixin.

This machinery is more simple now and easier to understand:
if you add a process/scenario
you have to explicitily pass the path for the demo resource.

Moreover, you can now pass the path for the service explicitily.
If not provided, it gets computed as a menu item path.

Finally, demo resources are loaded all at once as the app is loaded,
which fixes the flappy issue of not having all demo data available
immediately when switching process via UI.