front: fix event binding w/ wrapped state
Before this change the state events were bound using the current computed state object.
This was setting as handler an object that was bound to the OLD state
of the component which would cause `this.state` to be inconsistent
inside the handler.

Now the specification of each state is retrieved w/ proper api method
and the event is tied to the specification definition of the handler.

Furthermore, the event is unbind before binding it again.