delivery: fix failing test
When we call the "done" endpoint and a backorder is created, the test
ensure the backorder is out of scope: nothing in the "done" endpoint
ensures that the backorder is assigned; it calls
"picking.action_done()": this check is testing the behavior of
"action_done" which is supposed to be tested in the stock module.
Actually, this condition currently fails because of another module
(probably depends of the configuration), but this is not what we want
to test here.