[FIX] sf: fetch lot related to current product(s)
In Odoo a lot number can be shared by multiple lots as soon as the
product is different.
This has to be taken into account by Shopfloor each time we scan a lot
number by restricting the search to relevant products, or it could
result unexpected result (a "Barcode not found" for instance).