[IMP] shopfloor: compute package move_line_count when needed
If you want to get the data for a package it always computes the move_line_count.
This is very bad for perf on scenario - like zone picking - where you have a long list of moves with packages
w/o needing such information.

The only scenario which is using the move_line_count from the package data is the checkout.
So i think it is convenient to not compute it by default only if it is really needed.