[IMP] shopfloor, zone picking: Allow partial transfer when scanning a location.
In the zone picking scenario, operators can choose to pick less than the reserved quantity.
Previously, the system required them, in this case, to place the goods in a package first.

This restriction was unnecessary, as it forced operators to perform an extra step that they do not actually do in real life.
We trust them to know what they're doing, so let's keep it simple.

With this fix, if a user scans a location, they can move the products directly to the scanned location.